third_party/blink/web_tests/external/wpt/selection/textcontrols/selectionchange-on-shadow-dom-expected.txt

This is a testharness.js-based test.
[FAIL] selectionchange event fired on a shadow dom bubble to the document
  assert_equals: expected 1 but got 2
Harness: the test ran to completion.
